Move AsyncRead test utils to submodule
